fix: server doesn't close after tests complete
