+++ test/app.test.js
@@ -1,8 +1,7 @@
 require('dotenv').config();
 const http = require('http');
-const process = require('process');
-const myApp = require('../app.js');
-const request = require('supertest')(`http://localhost:3000/person`);
+const startServer = require('../src/server.js');
+const supertest = require('supertest');
 
 const createData = { name: 'Vadzim', age: 34, hobbies: ['computers'] };
 const updateData = {
@@ -12,16 +11,34 @@ const updateData = {
 };
 const badData = { name: 'Vadzim', age: 34, hobbies: 'programming' };
 
+// beforeAll(function () {
+server = startServer();
+request = supertest(server);
+// });
+
+// afterAll(function (done) {
+//   httpServ.close(done);
+// });
+
+// afterAll((done) => {
+//   server.close(done);
+// });
+
+afterAll(() => {
+  server.close();
+});
+
 describe('Hacker scope - E2E API Tests', function () {
   describe('First scenario - Read All => Create => Read Create => Update Created => Delete Updated => Read Deleted', function () {
     let personId;
 
     it('should return empty array of persons', async () => {
-      await request.get('/').expect(200, []);
+      await request.get('/person').expect(200, []);
     });
 
     it('should return created object of person with id', async () => {
-      const res = await request.post('/').send(createData);
+      const res = await request.post('/person').send(createData);
+      console.log(res.body);
       expect(res.status).toEqual(201);
       expect(res.body).toEqual(
         expect.objectContaining({ id: expect.any(String), ...createData })
@@ -30,7 +47,7 @@ describe('Hacker scope - E2E API Tests', function () {
     });
 
     it('should return object of person created in prev test', async () => {
-      const res = await request.get('/' + personId);
+      const res = await request.get('/person/' + personId);
       expect(res.status).toEqual(200);
       expect(res.body).toEqual(
         expect.objectContaining({ id: personId, ...createData })
@@ -38,7 +55,7 @@ describe('Hacker scope - E2E API Tests', function () {
     });
 
     it('should return updated object of person', async () => {
-      const res = await request.put('/' + personId).send(updateData);
+      const res = await request.put('/person/' + personId).send(updateData);
       expect(res.status).toEqual(200);
       expect(res.body).toEqual(
         expect.objectContaining({ id: personId, ...updateData })
@@ -46,13 +63,13 @@ describe('Hacker scope - E2E API Tests', function () {
     });
 
     it('should delete object of person by id and return 204', async () => {
-      const res = await request.delete('/' + personId);
+      const res = await request.delete('/person/' + personId);
       expect(res.status).toEqual(204);
       expect(res.body).toEqual('');
     });
 
     it('should return 404 and msg person is not found', async () => {
-      const res = await request.get('/' + personId);
+      const res = await request.get('/person/' + personId);
       expect(res.status).toEqual(404);
       expect(res.body).toEqual({
         error: `Person with id ${personId} not found!`,
@@ -64,7 +81,7 @@ describe('Hacker scope - E2E API Tests', function () {
     let personIds = [];
 
     it('should return internal error msg', async () => {
-      const res = await request.post('/').send(badData);
+      const res = await request.post('/person').send(badData);
       expect(res.status).toEqual(400);
       expect(res.body).toEqual({
         error: `Hobbies must be an array of strings or empty array!`,
@@ -75,7 +92,7 @@ describe('Hacker scope - E2E API Tests', function () {
       for (let i = 0; i < 3; i++) {
         const tmpCreateData = createData;
         tmpCreateData.hobbies.push(String(i + 1));
-        const res = await request.post('/').send(tmpCreateData);
+        const res = await request.post('/person').send(tmpCreateData);
         expect(res.status).toEqual(201);
         expect(res.body).toEqual(
           expect.objectContaining({ id: expect.any(String), ...tmpCreateData })
@@ -85,21 +102,21 @@ describe('Hacker scope - E2E API Tests', function () {
     });
 
     it('should return array of three persons', async () => {
-      const res = await request.get('/');
+      const res = await request.get('/person');
       expect(res.status).toEqual(200);
       expect(res.body.length).toEqual(3);
       // console.log(res.body);
     });
 
     it('should delete second person by id and return 204', async () => {
-      const res = await request.delete('/' + personIds[1]);
+      const res = await request.delete('/person/' + personIds[1]);
       personIds.splice(1, 1);
       expect(res.status).toEqual(204);
       expect(res.body).toEqual('');
     });
 
     it('should return array of two persons', async () => {
-      const res = await request.get('/');
+      const res = await request.get('/person');
       expect(res.status).toEqual(200);
       expect(res.body.length).toEqual(2);
       expect(res.body[1].hobbies).not.toEqual(['computers', '1', '2']);
@@ -107,7 +124,7 @@ describe('Hacker scope - E2E API Tests', function () {
 
     it('should delete two last persons', async () => {
       for (let i = 0; i < personIds.length; i++) {
-        const res = await request.delete('/' + personIds[i]);
+        const res = await request.delete('/person/' + personIds[i]);
         expect(res.status).toEqual(204);
         expect(res.body).toEqual('');
       }
@@ -118,7 +135,7 @@ describe('Hacker scope - E2E API Tests', function () {
     let duplicateData;
 
     it('should create new persons', async () => {
-      const res = await request.post('/').send(createData);
+      const res = await request.post('/person').send(createData);
       expect(res.status).toEqual(201);
       expect(res.body).toEqual(
         expect.objectContaining({ id: expect.any(String), ...createData })
@@ -128,14 +145,14 @@ describe('Hacker scope - E2E API Tests', function () {
     });
 
     it('should return error msg', async () => {
-      const res = await request.put('/' + duplicateData.id).send('');
+      const res = await request.put('/person/' + duplicateData.id).send('');
       console.log(res.body);
       expect(res.status).toEqual(400);
       expect(res.body).toEqual({ error: `Request body is missing!` });
     });
 
     it('should duplicate person with another id', async () => {
-      const res = await request.post('/').send(duplicateData);
+      const res = await request.post('/person').send(duplicateData);
       expect(res.status).toEqual(201);
       expect(res.body.id).not.toEqual(duplicateData.id);
       expect(res.body).toEqual(
@@ -149,7 +166,7 @@ describe('Hacker scope - E2E API Tests', function () {
     });
 
     it('should return array of two same persons with different ids', async () => {
-      const res = await request.get('/');
+      const res = await request.get('/person');
       expect(res.status).toEqual(200);
       expect(res.body.length).toEqual(2);
       expect(res.body[0].id).not.toEqual(res.body[1].id);
